public/assets/python/requirements.txt
paramiko==2.12.0
cryptography==40.0.2
pymysql==1.1.0
python-dotenv==1.0.1
pycryptodome==3.20.0
